requirements.txt
baidu-aip==4.16.13
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
darkdetect==0.8.0
idna==3.7
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pycparser==2.22
PyMySQL==1.1.0
PyQt-Fluent-Widgets==1.5.5
PyQt5==5.15.9
PyQt5-Frameless-Window==0.3.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
python-dotenv==1.0.1
pywin32==306
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
sip==6.8.3
tomli==2.0.1
urllib3==2.2.1
